refactor(event): 패키지 변경
* 이벤트가 증정, 할인 두 종류로 나뉘기 때문에 패키지 분리
