+++ src/main/java/christmas/domain/event/discount/HolidayDiscountEvent.java
@@ -1,4 +1,4 @@
-package christmas.domain.event;
+package christmas.domain.event.discount;
 
 import christmas.domain.Day;
 import christmas.domain.MenuType;
